some refactor

+++ src/main/java/ru/npcric/asparagus/trainerslog/service/TicketService.java
@@ -8,11 +8,13 @@ import org.springframework.stereotype.Service;
 import ru.npcric.asparagus.trainerslog.adapter.repository.StudentRepository;
 import ru.npcric.asparagus.trainerslog.adapter.repository.TicketRepository;
 import ru.npcric.asparagus.trainerslog.adapter.web.dto.request.ticket.TicketValidationRequest;
+import ru.npcric.asparagus.trainerslog.adapter.web.errors.UserNotFoundException;
 import ru.npcric.asparagus.trainerslog.domain.StudentEntity;
 import ru.npcric.asparagus.trainerslog.domain.TicketEntity;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.Optional;
 
 @Service
 @FieldDefaults(makeFinal = true, level = AccessLevel.PUBLIC)
@@ -21,10 +23,11 @@ public class TicketService {
     TicketRepository ticketRepository;
     StudentRepository studentRepository;
     int INITIAL_COST = 1000;
-    //метод для шедулера --> обнуляет тикеты у всех и создает все новый дефолт тикет
     @Transactional
     public void updateValidTicketToDefault(TicketValidationRequest ticketValidationRequest) {
-        StudentEntity studentEntity = studentRepository.findByUser_Username(ticketValidationRequest.username());
+        Optional<StudentEntity> student = studentRepository.findByUser_Username(ticketValidationRequest.username());
+        if (student.isEmpty()) throw new UserNotFoundException(ticketValidationRequest.username());
+        StudentEntity studentEntity = student.get();
         Long ticketId = studentEntity.getTicket().getId();
         studentRepository.updateStudentByIdSetTicketToNull(studentEntity.getId());
         ticketRepository.deleteById(ticketId);
@@ -35,7 +38,9 @@ public class TicketService {
 
     @Transactional
     public void updateToPaidTicket(TicketValidationRequest ticketValidationRequest) {
-        StudentEntity studentEntity = studentRepository.findByUser_Username(ticketValidationRequest.username());
+        Optional<StudentEntity> student = studentRepository.findByUser_Username(ticketValidationRequest.username());
+        if (student.isEmpty()) throw new UserNotFoundException(ticketValidationRequest.username());
+        StudentEntity studentEntity = student.get();
         TicketEntity ticket = studentEntity.getTicket();
         ticket.setPaidAmount(0);
         ticket.setIsExpired(false);
@@ -43,13 +48,11 @@ public class TicketService {
     }
 
     public TicketEntity createTicketForNewStudent() {
-        TicketEntity ticketEntity = getDefaultTicket();
-        return ticketEntity;
+        return getDefaultTicket();
     }
 
     public TicketEntity getDefaultTicket() {
-        TicketEntity ticketEntity = ticketRepository.save(new TicketEntity(LocalDate.now(),
+        return ticketRepository.save(new TicketEntity(LocalDate.now(),
                 LocalDate.now().plusMonths(1), true, INITIAL_COST));
-        return ticketEntity;
     }
 }
